refactor: remove repository layer after agreement with Sander
