+++ types/records.d.ts
@@ -70,53 +70,3 @@ interface RecordDayStatus {
   MID: string;
   LONG: string;
 }
-
-interface ITimeRecordsRepository {
-  getEmployeeRecords(
-    params: {
-      employeeId: string;
-      startDate?: string;
-      endDate?: string;
-    },
-    collection: string
-  ): Promise<RecordType[]>;
-  getRecords(params: {startDate: Date; endDate: Date}, collection: string): Promise<RecordType[]>;
-  addBridgeWorklogs(params: {
-    employeeId: string;
-    bridgeUid: string;
-    timeRecords: TimeRecord[];
-    contracts: number[];
-  }): Promise<void>;
-  addBridgeWorklog(
-    record: TimeRecord,
-    contractId: number,
-    bridgeUid: string,
-    ref: any
-  ): Promise<void>;
-  removeBridgeWorklogs(timeRecords: TimeRecord[]): Promise<void>;
-  removeBridgeWorklog(record: TimeRecord, ref: any): Promise<any>;
-  saveEmployeeRecords(
-    params: {
-      employeeId: string;
-      bridgeUid?: string;
-      timeRecords: RecordType[];
-      contracts?: number[];
-    },
-    collection: string
-  ): Promise<Awaited<RecordType>[]>;
-  deleteEmployeeRecords(params: {recordsToDelete: RecordType[]}, collection: string);
-  getWeeklyRecords(params: {employeeId: string; startDate: Date}): Promise<WeeklyRecords>;
-}
-
-interface ITravelRecordsRepository {
-  getEmployeeRecords(params: {
-    employeeId: string;
-    startDate?: string;
-    endDate?: string;
-  }): Promise<TravelRecord[]>;
-  getRecords(params: {startDate: Date; endDate: Date}): Promise<TimeRecord[]>;
-  saveEmployeeRecords(params: {
-    employeeId: string;
-    travelRecords: TravelRecord[];
-  }): Promise<TravelRecord[]>;
-}
